+++ LangGraph/workflow.py
@@ -10,47 +10,14 @@ import numpy as np
 from langchain_community.embeddings import HuggingFaceEmbeddings
 from langchain_community.vectorstores import FAISS
 
-
-CATEGORY_LIST = []
-
 # Load API Key
 load_dotenv()
-api_key = os.getenv("TOGETHER_API_KEY")
-
-# Initialize LLM
-llm = ChatOpenAI(
-    model="mistralai/Mistral-7B-Instruct-v0.2",
-    openai_api_base="https://api.together.xyz/v1",
-    openai_api_key=api_key,
-)
 
 # Define the state schema
 class QueryState(BaseModel):
     question: str
     category: str | None = None
     response: str | None = None
-
-# Define step 1: Categorizing the question
-def categorize_question(state: QueryState):
-    # Generate 10 categories
-    #Will eventually be a part of "Update" workflow
-    #NEED TO CHANGE THIS LINE LATER TO SET CATEGORY_LIST FROM DATABASE AND UPDATE CATEGORIES IN ANOTHER SCRIPT
-    CATEGORY_LIST = categories.main()
-
-    #predict is deprecated but this is likely going to end up in a script of its own anyway so we can fix it then
-    categories_str = ", ".join(CATEGORY_LIST)
-
-    # Compose a prompt that includes the list of available categories and the question
-    prompt = (
-        f"Given the following categories: {categories_str}. "
-        f"Categorize the following question: \"{state.question}\". "
-        f"Respond with only the category name that best fits the question."
-    )
-
-    # Use the LLM to determine the category
-    category = llm.predict(prompt).strip()
-    return QueryState(question=state.question, category=category)
-
 
 # Define step 2: Store embeddings (Placeholder) --> Will move into "UPDATE" workflow and store in Chroma rather than local vector store FAISS
 def create_and_store_embedding(state: QueryState):
