+++ requirements.txt
@@ -59,7 +59,7 @@ monotonic==1.6
 mpmath==1.3.0
 msgpack==1.1.0
 multidict==6.1.0
-numpy==1.26.4
+# numpy==1.26.4
 numpy==2.2.3
 oauthlib==3.2.2
 onnxruntime==1.20.1
